refactor: change i18n wording
