+++ src/components/dataverse/dataversePreview/dataversePreview.scss
@@ -4,21 +4,15 @@
   gap: 10px;
 
   h2 {
+    text-transform: capitalize;
     margin: 0px;
   }
 
   span {
-    &:first-child {
-      text-transform: uppercase;
-    }
-
-    &:nth-child(2),
-    &:nth-child(3),
-    &:nth-child(4) {
-      text-transform: capitalize;
-    }
+    text-transform: capitalize;
 
     &:nth-child(5) {
+      text-transform: none;
       text-align: end;
     }
   }
