+++ src/components/goToPreviousPage/GoToPreviousPage.tsx
@@ -14,7 +14,7 @@ export const GoToPreviousPage = (): JSX.Element => {
 
   return (
     <div className="okp4-go-to-previous-page">
-      <Button label={t("explore:back")} leftIcon={<Icon name="arrow-left" size={22} />} onClick={handleClick} />
+      <Button label={t("dataverse:back")} leftIcon={<Icon name="arrow-left" size={22} />} onClick={handleClick} />
     </div>
   )
 }
