+++ src/i18n/dataverse/dataverse_en.json
@@ -1,0 +1,21 @@
+{
+  "addThis": "Add this",
+  "back": "Back",
+  "by": "By",
+  "completude": "Completude",
+  "dataset": "Dataset",
+  "dataspace": "Dataspace",
+  "description": "Description",
+  "format": "Format",
+  "jointure": "Jointure",
+  "metadata": "Metadata",
+  "private": "Private",
+  "provider": "Provider",
+  "public": "Public",
+  "quality": "Quality",
+  "service": "Service",
+  "size": "Size",
+  "title": "Explore the Dataverse",
+  "toDataspace": "to my Dataspace",
+  "updatedAt": "Last update"
+}
