--- src/i18n/explore/explore_en.json
@@ -1,18 +1,0 @@
-{
-  "by": "By {{provider}} - Last update {{- updated}}",
-  "list": {
-    "configuration": {
-      "show": "Show:",
-      "elements": "elements",
-      "display": "Display:",
-      "grid": "Grid",
-      "list": "List",
-      "sort": "Sort:",
-      "select": {
-        "alphabetical": "Alphabetical",
-        "created-on": "Date of creation"
-      }
-    }
-  },
-  "title": "Explore the Dataverse"
-}
