--- src/i18n/explore/explore_fr.json
@@ -1,18 +1,0 @@
-{
-  "by": "Par {{provider}} - Dernière modification {{- updated}}",
-  "list": {
-    "configuration": {
-      "show": "Lister:",
-      "elements": "éléments",
-      "display": "Afficher:",
-      "grid": "Grille",
-      "list": "Liste",
-      "sort": "Trier par:",
-      "select": {
-        "alphabetical": "Alphabétique",
-        "created-on": "Date de création"
-      }
-    }
-  },
-  "title": "Explorez le Dataverse"
-}
