+++ src/i18n/governance/governance_en.json
@@ -1,0 +1,5 @@
+{
+  "name": "Governance",
+  "view": "View Governance",
+  "based": "Governance based on the $KNOW token"
+}
