+++ src/i18n/governance/governance_fr.json
@@ -1,0 +1,5 @@
+{
+  "name": "Gouvernance",
+  "view": "Voir la gouvernance",
+  "based": "Gouvernance basée sur le $KNOW token"
+}
