+++ src/pages/dataverses/dataverseId.scss
@@ -47,7 +47,7 @@
     grid-area: information;
 
     .okp4-portal-card {
-      border-radius: 0px 0px 0px 9px;
+      border-radius: 0px 0px 9px 9px;
     }
   }
 }
